funman state to track inprogress simulation to allow better resilience
